# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Civix API - civic issue reporting and community engagement platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: issues
    description: Civic issue reports
  - name: polls
    description: Voting system polls
  - name: feedback
    description: User feedback
  - name: contact
    description: Contact messages
  - name: contributors
    description: Contributors leaderboard
  - name: sos
    description: SOS emergency alerts

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /issues:
    get:
      operationId: getIssues
      tags: [issues]
      summary: List all civic issues
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Report a new civic issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueBody"
      responses:
        "201":
          description: Issue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"

  /issues/stats:
    get:
      operationId: getIssueStats
      tags: [issues]
      summary: Get issue statistics summary
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueStats"

  /issues/{id}:
    get:
      operationId: getIssue
      tags: [issues]
      summary: Get a single issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateIssue
      tags: [issues]
      summary: Update an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
    delete:
      operationId: deleteIssue
      tags: [issues]
      summary: Delete an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /issues/{id}/upvote:
    post:
      operationId: upvoteIssue
      tags: [issues]
      summary: Upvote an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upvoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"

  /polls:
    get:
      operationId: getPolls
      tags: [polls]
      summary: List polls
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poll"
    post:
      operationId: createPoll
      tags: [polls]
      summary: Create a poll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePollBody"
      responses:
        "201":
          description: Poll created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"

  /polls/stats:
    get:
      operationId: getPollStats
      tags: [polls]
      summary: Get voting system stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollStats"

  /polls/{id}:
    get:
      operationId: getPoll
      tags: [polls]
      summary: Get a poll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"

  /polls/{id}/vote:
    post:
      operationId: voteOnPoll
      tags: [polls]
      summary: Vote on a poll option
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteBody"
      responses:
        "200":
          description: Voted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"

  /feedback:
    get:
      operationId: getFeedback
      tags: [feedback]
      summary: Get all feedback (admin)
      responses:
        "200":
          description: List of feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitFeedbackBody"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"

  /contact:
    post:
      operationId: sendContactMessage
      tags: [contact]
      summary: Send a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /contributors:
    get:
      operationId: getContributors
      tags: [contributors]
      summary: Get leaderboard of contributors
      parameters:
        - name: sort
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Contributors list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContributorsResponse"

  /contributors/{username}:
    get:
      operationId: getContributor
      tags: [contributors]
      summary: Get a contributor by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contributor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contributor"

  /sos:
    post:
      operationId: triggerSOS
      tags: [sos]
      summary: Trigger an SOS emergency alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SOSBody"
      responses:
        "201":
          description: SOS triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SOSResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    RegisterBody:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        name:
          type: string
        location:
          type: string

    LoginBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        bio:
          type: string
        avatar:
          type: string

    User:
      type: object
      required: [id, username, email, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        role:
          type: string
        issueCount:
          type: integer
        createdAt:
          type: string

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    Issue:
      type: object
      required: [id, title, category, status, location, upvotes, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        location:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        upvotes:
          type: integer
        imageUrl:
          type: ["string", "null"]
        reportedBy:
          type: ["string", "null"]
        reporterName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    CreateIssueBody:
      type: object
      required: [title, category, location]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        lat:
          type: number
        lng:
          type: number
        imageUrl:
          type: string

    UpdateIssueBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string

    IssueStats:
      type: object
      required: [total, open, inProgress, resolved, categories]
      properties:
        total:
          type: integer
        open:
          type: integer
        inProgress:
          type: integer
        resolved:
          type: integer
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"

    CategoryCount:
      type: object
      required: [name, count]
      properties:
        name:
          type: string
        count:
          type: integer

    Poll:
      type: object
      required: [id, question, options, totalVotes, status, createdAt]
      properties:
        id:
          type: string
        question:
          type: string
        description:
          type: ["string", "null"]
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        totalVotes:
          type: integer
        status:
          type: string
        createdBy:
          type: ["string", "null"]
        creatorName:
          type: ["string", "null"]
        endsAt:
          type: ["string", "null"]
        createdAt:
          type: string

    PollOption:
      type: object
      required: [id, text, votes]
      properties:
        id:
          type: string
        text:
          type: string
        votes:
          type: integer

    CreatePollBody:
      type: object
      required: [question, options]
      properties:
        question:
          type: string
        description:
          type: string
        options:
          type: array
          items:
            type: string
        endsAt:
          type: string

    VoteBody:
      type: object
      required: [optionId]
      properties:
        optionId:
          type: string

    PollStats:
      type: object
      required: [activePolls, totalVotes, engagement]
      properties:
        activePolls:
          type: integer
        totalVotes:
          type: integer
        engagement:
          type: integer

    Feedback:
      type: object
      required: [id, category, rating, message, createdAt]
      properties:
        id:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        category:
          type: string
        rating:
          type: integer
        message:
          type: string
        createdAt:
          type: string

    SubmitFeedbackBody:
      type: object
      required: [category, rating, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        category:
          type: string
        rating:
          type: integer
        message:
          type: string

    ContactMessageBody:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
        email:
          type: string
        message:
          type: string

    Contributor:
      type: object
      required: [id, username, contributions, rank, badge]
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        contributions:
          type: integer
        rank:
          type: integer
        badge:
          type: string
        location:
          type: ["string", "null"]
        joinedAt:
          type: string

    ContributorsResponse:
      type: object
      required: [contributors, stats]
      properties:
        contributors:
          type: array
          items:
            $ref: "#/components/schemas/Contributor"
        stats:
          $ref: "#/components/schemas/ContributorStats"

    ContributorStats:
      type: object
      required: [totalContributors, totalContributions, topContributions]
      properties:
        totalContributors:
          type: integer
        totalContributions:
          type: integer
        topContributions:
          type: integer

    SOSBody:
      type: object
      required: [location, description]
      properties:
        location:
          type: string
        description:
          type: string
        lat:
          type: number
        lng:
          type: number
        contactName:
          type: string

    SOSResponse:
      type: object
      required: [id, status, message]
      properties:
        id:
          type: string
        status:
          type: string
        message:
          type: string